alertprocessor/pyproject.toml
[project]
name = "alerthandler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "firebase-admin>=7.0.0",
    "google-cloud-pubsub>=2.31.0",
]
